Extract debug routes
